backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
supabase==2.3.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.0
redis==5.0.0
openclaw==0.5.0
mcp==0.1.0
resend==0.8.0
pytest==7.4.0
pytest-asyncio==0.23.0
